Extend 'set_blink_rate()' to turn off display with value of '-1'
I have a project with multiple modules, when 'blink()' was enabled
they blinked a slightly different rates... which looked funky.

The register for blinking also includes a 'display_on' bit, so I
access that with a vaule of '-1' to turn off the display(s). Set a
normal blink rate to turn display(s) back on.

Tested on the 'HT16K33Segment' from Adafruit.